Mejora en arquitectura, eliminar LoginScreen mal generada
