+++ app/src/main/java/com/example/deliverysya/ui/navigation/AppScreen.kt
@@ -3,7 +3,6 @@ package com.example.deliverysya.ui.navigation
 sealed class AppScreen (val route: String){
     object SplashScreen: AppScreen("splash_screen")
     object LoginScreen: AppScreen("login_screen")
-    object LoginScreen2: AppScreen("login_screen2")
     object IntroductionRiders: AppScreen("introduction_riders")
 
 }
