+++ app/src/main/java/com/example/deliverysya/ui/screens/SplashScreen.kt
@@ -7,26 +7,20 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.material.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.example.deliverysya.R
 import com.example.deliverysya.ui.navigation.AppScreen
 import com.example.uicomponents.TitleText
 import kotlinx.coroutines.delay
-import kotlin.time.times
 
 @Composable
 fun SplashScreen(modifier: Modifier = Modifier, navController: NavController) {
@@ -35,7 +29,7 @@ Column() {
     LaunchedEffect(key1 = true) {
         delay(timeMillis)
         navController.popBackStack()
-        navController.navigate(AppScreen.LoginScreen2.route)
+        navController.navigate(AppScreen.LoginScreen.route)
     }
     SplashScreenInitialApp(
         modifier = modifier
